+++ test/data.test.ts
@@ -5,13 +5,15 @@ import { TSTest } from '../src/tstester';
 
 export const dataGroups = [
     TSTest.group("Testing simple level data manipulations functions", async (group) => {
-        const a:Uint8Array = new Uint8Array([65,66,67,68]) ;
+        const a = new ArrayBuffer(4);
+        const view = new Uint8Array(a);
+        view[0] = 65; view[1] = 66; view[2] = 67; view[3] = 68;
         const c = Buffer.from("ABCD") ;
         const b64 = 'JVBERi0xLjQKJcKlwrEKCgoKMSAwIG9iagogIDw8IC9UeXBlIC9DYXRhbG9nCiAgICAgL1BhZ2VzIDIgMCBSCiAgPj4KZW5kb2JqCgoyIDAgb2JqCiAgPDwgL1R5cGUgL1BhZ2VzCiAgICAgL0tpZHMgWzMgMCBSXQogICAgIC9Db3VudCAxCiAgICAgL01lZGlhQm94IFswIDAgMzAwIDE0NF0KICA+PgplbmRvYmoKCjMgMCBvYmoKICA8PCAgL1R5cGUgL1BhZ2UKICAgICAgL1BhcmVudCAyIDAgUgogICAgICAvUmVzb3VyY2VzCiAgICAgICA8PCAvRm9udAogICAgICAgICAgIDw8IC9GMQogICAgICAgICAgICAgICA8PCAvVHlwZSAvRm9udAogICAgICAgICAgICAgICAgICAvU3VidHlwZSAvVHlwZTEKICAgICAgICAgICAgICAgICAgL0Jhc2VGb250IC9UaW1lcy1Sb21hbgogICAgICAgICAgICAgICA+PgogICAgICAgICAgID4+CiAgICAgICA+PgogICAgICAvQ29udGVudHMgNCAwIFIKICA+PgplbmRvYmoKCjQgMCBvYmoKICA8PCAvTGVuZ3RoIDU1ID4+CnN0cmVhbQogIEJUCiAgICAvRjEgMTggVGYKICAgIDAgMCBUZAogICAgKEhlbGxvIFdvcmxkKSBUagogIEVUCmVuZHN0cmVhbQplbmRvYmoKCnhyZWYKMCA1CjAwMDAwMDAwMDAgNjU1MzUgZiAKMDAwMDAwMDAxOCAwMDAwMCBuIAowMDAwMDAwMDc3IDAwMDAwIG4gCjAwMDAwMDAxNzggMDAwMDAgbiAKMDAwMDAwMDQ1NyAwMDAwMCBuIAp0cmFpbGVyCiAgPDwgIC9Sb290IDEgMCBSCiAgICAgIC9TaXplIDUKICA+PgpzdGFydHhyZWYKNTY1CiUlRU9GCg==' ;        
         const u8array = $decodeBase64(b64) ;
 
         group.unary('x.isGenuineUint8Array() method', async(t) => {
-            t.expect0(a.isGenuineUint8Array()).true() ;
+            t.expect0(view.isGenuineUint8Array()).true() ;
             t.expect1(c.isGenuineUint8Array()).false() ;
         }) ;
         group.unary('$bufferFromArrayBuffer() and $arrayBufferFromBytes() functions', async(t) => {
@@ -22,9 +24,9 @@ export const dataGroups = [
         }) ;
 
         group.unary('$bufferFromBytes(), $uint8ArrayFromBytes(), $arrayFromBytes() functions', async(t) => {
-            const buf = $bufferFromBytes(a) ;
-            const u8a = $uint8ArrayFromBytes(a, { forceCopy:true }) ;
-            const bytes = $arrayFromBytes(a) ;
+            const buf = $bufferFromBytes(view) ;
+            const u8a = $uint8ArrayFromBytes(view, { forceCopy:true }) ;
+            const bytes = $arrayFromBytes(view) ;
 
             t.expect0(buf).is(a) ;
             t.expect1(u8a).is(a) ;
@@ -47,18 +49,18 @@ export const dataGroups = [
             const str64 = $encodeBase64(str) ;
             t.expect2(str64).is(Buffer.from(str, 'binary').toString('base64')) ;
             t.expect3(b64_2.decodeBase64()).is(u8array) ;
-            t.expect4(b64_2.decodeBase64().toBase64()).is(b64_2) ;
-            t.expect5(u8array.toBase64().decodeBase64()).is(u8array) ;
+            t.expect4(b64_2.decodeBase64().base64String()).is(b64_2) ;
+            t.expect5(u8array.base64String().decodeBase64()).is(u8array) ;
 
-            let utf64 = utf8String.toBase64() ;
+            let utf64 = utf8String.base64String() ;
             t.expect6(utf64).isnot(utf8Str64) ;
-            utf64 = utf8String.toBase64(TSCharset.utf8Charset()) ;
+            utf64 = utf8String.base64String(TSCharset.utf8Charset()) ;
             t.expect7(utf64).is(utf8Str64) ;
             t.expect8($decodeBase64(utf64)).is(uft8Array) ;
 
-            utf64 = utf8String.toBase64URL() ;
+            utf64 = utf8String.base64URL() ;
             t.expect9(utf64).isnot(utf8Str64URL) ;
-            utf64 = utf8String.toBase64URL(TSCharset.utf8Charset()) ;
+            utf64 = utf8String.base64URL(TSCharset.utf8Charset()) ;
             t.expectA(utf64).is(utf8Str64URL) ;
             const durl64 = $decodeBase64URL(utf64) ;
             t.expectB(TSCharset.utf8Charset().stringFromData(durl64)).is(utf8String) ;